+++ src/lib.rs
@@ -35,10 +35,10 @@
 //!
 //! // Your game logic functions can be named anything, but the first parameter is always a
 //! // `&mut EngineState`, and the second parameter is a mutable reference to your custom game
-//! // state struct (`&mut GameState` in this case). The function returns a `bool`.
+//! // state struct (`&mut GameState` in this case).
 //! //
 //! // This function will be run once each frame.
-//! fn game_logic(engine_state: &mut EngineState, game_state: &mut GameState) -> bool {
+//! fn game_logic(engine_state: &mut EngineState, game_state: &mut GameState) {
 //!     // The `EngineState` contains all sorts of built-in goodies.
 //!     // Get access to the player sprite...
 //!     let player = engine_state.sprites.get_mut("player").unwrap();
@@ -48,8 +48,6 @@
 //!     if player.translation.x > 100.0 {
 //!         game_state.health -= 1;
 //!     }
-//!     // Returning `true` means the next logic function in line should be run.
-//!     true
 //! }
 //! ```
 //!